pyproject.toml
[project]
name = "etl-pyspark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyspark>=4.0.1",
    "pyyaml>=6.0.3",
    "unidecode>=1.4.0",
]
